refactor: split assertions and mocks
